Yes, It's back!

+++ src/main/java/sh/talonfox/ravenstone/processor/ProcessorHost.java
@@ -3,8 +3,8 @@ package sh.talonfox.ravenstone.processor;
 public interface ProcessorHost {
     byte busRead(byte id, byte at);
     void busWrite(byte id, byte at, byte val);
-    byte memRead(int at);
-    void memStore(int at, byte data);
+    byte memRead(long at);
+    void memStore(long at, byte data);
     void invalidatePeripheral();
 
     void stop();
